requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
flask-cors==4.0.0
supabase==2.3.4
openai==1.12.0
quart==0.19.4
quart-cors==0.7.0
python-multipart==0.0.9
hypercorn==0.16.0 